Move 44bbp compression into generic bitpacking function. Align all sprite data to u32.

+++ GbaGameEngine/src/engine/gba/graphics/tiles/GBATileBank.h
@@ -19,7 +19,7 @@ namespace GBA
 		BlockGroupCount
 	};
 
-	typedef List<u16> tSpriteData;
+	typedef List<u32> tSpriteData;
 	
 	class TileBank
 	{
@@ -39,11 +39,11 @@ namespace GBA
 		static volatile CharBlock8* EditTileBlock8(TileBlockGroups group) { return &(s_charBlockPool8[int(group)]); }
 
 	public:
-		static bool LoadTiles(const u16* pixelMap, u32 pixelMapSize, TileBlockGroups tileBlockGroup, u16 startTileIndex);
-		static bool LoadTiles(const List<u16>& pixelMap, TileBlockGroups tileBlockGroup, u16 startTileIndex);
+		static bool LoadTiles(const u32* pixelMap, u32 pixelMapSize, TileBlockGroups tileBlockGroup, u16 startTileIndex);
+		static bool LoadTiles(const List<u32>& pixelMap, TileBlockGroups tileBlockGroup, u16 startTileIndex);
 
 		template <u32 SIZE>
-		static bool LoadTiles(const Array<u16, SIZE>& pixelMap, TileBlockGroups tileBlockGroup, u16 startTileIndex)
+		static bool LoadTiles(const Array<u32, SIZE>& pixelMap, TileBlockGroups tileBlockGroup, u16 startTileIndex)
 		{
 			return LoadTiles(pixelMap.GetContainer(), pixelMap.Count(), tileBlockGroup, startTileIndex);
 		}
